Alterado preco do produto de Double para BigDecimal

+++ src/main/java/br/com/fiap/techchallenge01/produto/domain/dto/request/ProdutoRequestDTO.java
@@ -1,9 +1,10 @@
 package br.com.fiap.techchallenge01.produto.domain.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 public class ProdutoRequestDTO {
@@ -17,6 +18,6 @@ public class ProdutoRequestDTO {
     private String idCategoria;
 
     @NotNull
-    private Double preco;
+    private BigDecimal preco;
 
 }
